test_pipeline/requirements.txt
albumentations==0.5.2
certifi==2020.11.8
chardet==3.0.4
click==7.1.2
configparser==5.0.1
cycler==0.10.0
decorator==4.4.2
docker-pycreds==0.4.0
docopt==0.6.2
gitdb==4.0.5
GitPython==3.1.11
idna==2.10
imageio==2.9.0
imgaug==0.4.0
kiwisolver==1.3.1
matplotlib==3.3.3
mkl-fft
mkl-random
networkx==2.5
numpy
olefile==0.46
opencv-python==4.4.0.46
opencv-python-headless==4.4.0.46
pathtools==0.1.2
Pillow 
promise==2.3
protobuf==3.14.0
psutil==5.7.3
pyparsing==2.4.7
python-dateutil==2.8.1
PyWavelets==1.1.1
PyYAML==5.3.1
requests==2.25.0
scikit-image==0.17.2
scipy==1.5.4
sentry-sdk==0.19.4
Shapely==1.7.1
shortuuid==1.0.1
six 
smmap==3.0.4
subprocess32==3.5.4
tifffile
torch
torchvision
tqdm==4.54.0
typing-extensions
urllib3==1.26.2
wandb==0.10.11
watchdog==0.10.4
wincertstore==0.2
yacs==0.1.8
yarg==0.1.9
